requirements.txt
fastapi==0.115.5
uvicorn==0.30.6
pymongo==4.6.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
email-validator==2.1.0.post1
python-dotenv==1.0.1
pydantic==2.9.2
annotated-types==0.6.0
typing-extensions==4.12.2
bcrypt==4.1.2
